test(web): split schedules page suites [OPE-392]
